+++ repositories/myPage.repository.js
@@ -16,27 +16,6 @@ const {
 class MypageRepository {
     constructor() { }
 
-    findUserById = async ({ userId }) => {
-        return await Users.findOne({
-            raw: true,
-            where: { userId },
-            attributes: [
-                "userId",
-                "userName",
-                "Team.teamName",
-                "remainDay",
-                "salaryDay",
-                "profileImg"
-            ],
-            include: [
-                {
-                    model: Teams,
-                    attributes: [],
-                },
-            ],
-        });
-    };
-
     getUserSchedule = async ({ userId }) => {
         const schedules = await Schedules.findAll({
             raw: true,
