+++ repositories/userInfo.repository.js
@@ -1,0 +1,27 @@
+const {Users,Teams} = require("../models")
+class UserInfoRepository {
+    constructor(){}
+
+    findUserById = async ({ userId }) => {
+        return await Users.findOne({
+            raw: true,
+            where: { userId },
+            attributes: [
+                "userId",
+                "userName",
+                "Team.teamName",
+                "remainDay",
+                "salaryDay",
+                "profileImg"
+            ],
+            include: [
+                {
+                    model: Teams,
+                    attributes: [],
+                },
+            ],
+        });
+    };
+}
+
+module.exports=UserInfoRepository;
