+++ routes/myPage.route.js
@@ -5,10 +5,6 @@ const mypagecontroller = new MypageController();
 const authmiddleware = require("../middlewares/auth-middleware.js");
 const router = express.Router();
 
-//유저정보조회
-//localhost:3003/usersInfo
-router.get("/usersInfo", authmiddleware, mypagecontroller.getUserInfo);
-
 //내 결재 조회
 //localhost:3003/mySchedule
 router.get("/mySchedule", authmiddleware, mypagecontroller.getSchedules);
